+++ docs/changes.md
@@ -9,6 +9,7 @@ title: Memory Change Log
 
 * Replace `array_set()` with `array_add()` when loading value from storage.
 * Avoid caching fallback instance since this could cause multiple redundant insertion to the database.
+* Allow to get handler instance using `Orchestra\Memory\Provider::getHandler()`.
 
 ### v2.1.2 {#v2-1-2}